netlink: retry interrupted dump
When two netlink dumps are called concurrently, the dump can be
interrupted. This is indicated by nlmsg flag NLM_F_DUMP_INTR in one of
the headers, and the dump can be inconsistent.

Whenever this occurs, restart the dump. Limit the number of restarts.

Inconsistency can still happen when interfaces change in between dump
calls. This is not yet covered.

Signed-off-by: Ondrej Hlavaty <ohlavaty@redhat.com>