netlink: message buffer based API
Introduce struct nlmsg which is a message buffer. It allows data and
attributes to be pushed to the buffer tail (put) and popped off the buffer
head (get) and resizes itself automatically as needed.

To access the message header regardless of the current head position,
nlmsg_get_hdr macro is available.

nla_attrs is converted to nlmsg_attrs and accepts the message buffer.

Two functions that ease creating of rtnetlink and genetlink message are
provided. A helper function for rtnetlink ifinfo dump is provided, too; it
should be replaced by a more generic rtnl dump function in the future, if
possible.

Signed-off-by: Jiri Benc <jbenc@redhat.com>