netns: more robust handling of /var/run/netns/ files
It turns out that in some cases, there may be files left in /var/run/netns/
that are not valid netns mounts. Most likely it's a result of some software
incorrectly cleaning up or crashing between unmounting and removing a file
there.

Handle this gracefully by producing a warning to the output file rather than
terminating.

Reference: https://unix.stackexchange.com/questions/610466/is-there-a-tool-to-print-the-network-topology-structure-in-linux
Signed-off-by: Jiri Benc <jbenc@redhat.com>