add link-netnsid for devices without a peer
Some devices, such as VXLAN tunnels, export a IFLA_LINK_NETNSID, which
is the netns in which they lookup routes for the outer IP header, but
don't have a corresponding link (IFLA_LINK). Currently, we're not
dumping that information.

This patch adds a link_net relation to each if_entry, which is set
similarly to the way peers are found. This is then printed by the
frontends only when the device doesn't have a link.

In the dot frontend, we add a dashed arrow (like for IF_LINK_WEAK
links) to the box of the peer netns, or to an unboxed "root netns"
label in the case of init_net. This label is only shown when needed,
so output_ifaces_pass1 is updated to check if we need it.

In the json frontend, links that don't already have a peer or
parent/child relationship now get a "link-netns" attribute, containing
the "id" attribute of the namespace.

Signed-off-by: Sabrina Dubroca <sd@queasysnail.net>